Some formatting...
git-svn-id: https://svn.apache.org/repos/asf/avalon/trunk@45947 13f79535-47bb-0310-9956-ffa450edef68

+++ central/laboratory/studio/core/templates/context/custom/NumberCruncher.java
@@ -26,7 +26,7 @@ public class NumberCruncher
     private final int m_primary;
     private final float m_secondary;
 
-    public NumberCruncher( Integer primary, Double secondary )
+    public NumberCruncher( Integer primary, Float secondary )
     {
         m_primary = primary.intValue();
         m_secondary = secondary.floatValue();
